market_risk_premiums

FinancialModelingPrep.market_risk_premiums — Function

market_risk_premiums(fmp)

Returns a vector of market risk premiums for each country. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Market-Risk-Premium for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all market risk premiums
data = market_risk_premiums(fmp)

treasury_rates

FinancialModelingPrep.treasury_rates — Function

treasury_rates(fmp, params...)

Returns a list of the treasury rates for the specified date range. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Treasury-Rates for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the treasury rates for Q1 2022
data = treasury_rates(fmp, from = "2022-01-01", to = "2022-03-31")

economic_indicator

FinancialModelingPrep.economic_indicator — Function

economic_indicator(fmp, name, params...)

Returns a list of the treasury rates for the specified date range. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: An econmic indicator.

params...: Additional keyword query params.

See Economic-Indicator for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the real GDP for Q1 2022
data = economic_indicator(fmp, name = "realGDP", from = "2022-01-01", to = "2022-03-31")

# get the federal funds rate for Q1 2022
data = economic_indicator(fmp, name = "federalFunds", from = "2022-01-01", to = "2022-03-31")