insider_trading_types

FinancialModelingPrep.insider_trading_types — Function

insider_trading_types(fmp)

Returns insider trading transaction types.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Insider-Trading for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the types of insider trading events
data = insider_trading_types(fmp)

insider_trades

FinancialModelingPrep.insider_trades — Function

insider_trades(fmp, params...)

Returns insider trades.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Insider-Trading for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the most recent insider purchases and sales
data = insider_trades(fmp, transactionType = ["P-Purchase", "S-Sale"], page = 0)

# get the most recent insider trades for AAPL
data = insider_trades(fmp, symbol = "AAPL", page = 0)

insider_trades_feed

FinancialModelingPrep.insider_trades_feed — Function

insider_trades_feed(fmp, params...)

Returns insider trades from the RSS feed.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Insider-Trading-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the most recent insider trades from the feed
data = insider_trades_feed(fmp, page = 0)

insiders_list

FinancialModelingPrep.insiders_list — Function

insiders_list(fmp, params...)

Returns all insider names and their CIKs matching the specified parameters.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See CIK-Mapper for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all insider names and their CIKs from page 3
data = insiders_list(fmp, page = 3)

cik_from_insider

FinancialModelingPrep.cik_from_insider — Function

cik_from_insider(fmp, name)

Returns all CIKs matching the specified insider name.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: A name.

See CIK-Mapper for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all CIKs matching "zuckerberg mark"
data = cik_from_insider(fmp, "zuckerberg%20mark")

cik_from_symbol

FinancialModelingPrep.cik_from_symbol — Function

cik_from_symbol(fmp, symbol)

Returns all CIKs matching the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See CIK-Mapper for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the CIK for AAPL
data = cik_from_symbol(fmp, "AAPL")

insider_roster

FinancialModelingPrep.insider_roster — Function

insider_roster(fmp, symbol)

Returns a JSON table with the insider roster for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Insider-Roster for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the insider roster for AAPL
data = insider_roster(fmp, "AAPL")

insider_roster_statistics

FinancialModelingPrep.insider_roster_statistics — Function

insider_roster_statistics(fmp, symbol)

Returns insider roster statistics for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Insider-Roster-Statistics for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the insider roster statistics for AAPL
data = insider_roster_statistics(fmp, "AAPL")

fails_to_deliver

FinancialModelingPrep.fails_to_deliver — Function

fails_to_deliver(fmp, symbol, params)

Returns fails to deliver for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Fails-to-Deliver for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of fails to deliver for AAPL
data = fails_to_deliver(fmp, "AAPL", page = 0)