institutional_positions

FinancialModelingPrep.institutional_positions — Function

institutional_positions(fmp, symbol, params...)

Returns institutional positions for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Institutional-Stock-Ownership for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the institutional positions for AAPL including the current quarter
data = institutional_positions(fmp, "AAPL", includeCurrentQuarter = true)

institutional_ownership_percentages

FinancialModelingPrep.institutional_ownership_percentages — Function

institutional_ownership_percentages(fmp, symbol, params...)

Returns institutional ownership percentages for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Stock-Ownership-by-Holders for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the institutional ownership percentages for AAPL in Q3 of 2021
data = institutional_ownership_percentages(fmp, "AAPL", date = "2021-09-30")

institutional_ownership_weightings

FinancialModelingPrep.institutional_ownership_weightings — Function

institutional_ownership_weightings(fmp, symbol, params...)

Returns institutional ownership weightings for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Institutional-Stock-by-Shares-Held-and-Date for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the institutional ownership weightings for AAPL in Q3 of 2021
data = institutional_ownership_weightings(fmp, "AAPL", date = "2021-09-30")

institutional_ownership_feed

FinancialModelingPrep.institutional_ownership_feed — Function

institutional_ownership_feed(fmp, params...)

Returns institutional ownership from the RSS feed.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Institutional-Holder-Rss-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of results from the RSS feed
data = institutional_ownership_feed(fmp, page = 0)

institution_search

FinancialModelingPrep.institution_search — Function

institution_search(fmp, name)

Returns all CIKs matching the specified institution name.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: An institution name.

See Institutional-Holders-Search for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the CIKs for Berkshire Hathaway
data = institution_search(fmp, name = "Berkshire%20Hathaway%20Inc")

institution_portfolio_dates

FinancialModelingPrep.institution_portfolio_dates — Function

institution_portfolio_dates(fmp, cik)

Returns the portfolio dates for the specified institution.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

cik::String: A CIK.

See Institutional-Holders-Available-Date for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the portfolio dates for Berkshire Hathaway
data = institution_portfolio_dates(fmp, cik = "0001067983")

institution_portfolio_summary

FinancialModelingPrep.institution_portfolio_summary — Function

institution_portfolio_summary(fmp, cik)

Returns the portfolio summary for the specified institution.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

cik::String: A CIK.

See Institutional-Holdings-Portfolio-Positions-Summary for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the portfolio summary for Berkshire Hathaway
data = institution_portfolio_summary(fmp, cik = "0001067983")

institution_portfolio_industry_summary

FinancialModelingPrep.institution_portfolio_industry_summary — Function

institution_portfolio_industry_summary(fmp, cik, params...)

Returns the portfolio industry summary for the specified institution.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

cik::String: A CIK.

params...: Additional keyword query params.

See Institutional-Holdings-Portfolio-Industry-Summary for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the portfolio industry summary for Berkshire Hathaway in Q3 of 2021
data = institution_portfolio_industry_summary(fmp, cik = "0001067983", date = "2021-09-30")

institution_portfolio_composition

FinancialModelingPrep.institution_portfolio_composition — Function

institution_portfolio_composition(fmp, cik, params...)

Returns the portfolio composition for the specified institution.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

cik::String: A CIK.

params...: Additional keyword query params.

See Institutional-Holdings-Portfolio-Composition for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the portfolio composition for Berkshire Hathaway in Q3 of 2021
data = institution_portfolio_composition(fmp, cik = "0001067983", date = "2021-09-30")