price_quote

FinancialModelingPrep.price_quote — Function

price_quote(fmp, symbol)

Returns the price quote for the specified symbol(s). Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A financial symbol.

See Stock-Quote for more details.
See Index-Quote for more details.
See Euronext-Quote for more details.
See TSX-Quote for more details.
See Crypto-Quote for more details.
See Forex-Quote for more details.
See Commodity-Quote for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price quote for AAPL
data = price_quote(fmp, "AAPL")

price_quotes

FinancialModelingPrep.price_quotes — Function

price_quotes(fmp, market)

Returns the price quotes for the specified marklet. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

market::String: A financial market.

See Stock-Quote for more details.
See Index-Quote for more details.
See Euronext-Quote for more details.
See TSX-Quote for more details.
See Crypto-Quote for more details.
See Forex-Quote for more details.
See Commodity-Quote for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price quotes for major indexes
data = price_quote(fmp, "index")

# get the price quotes for NYSE
data = price_quote(fmp, "nyse")

# get the price quotes for TSX
data = price_quote(fmp, "euronext")

# get the price quotes for euronext
data = price_quote(fmp, "tsx")

# get the price quotes for crypto
data = price_quote(fmp, "crypto")

# get the price quotes for forex
data = price_quote(fmp, "forex")

# get the price quotes for commodity
data = price_quote(fmp, "commodity")

historical_price_quote

FinancialModelingPrep.historical_price_quote — Function

historical_price_quote(fmp, symbol, frequency = TIME_FREQUENCIES.daily, params...)

Returns the historical price quote for the specified symbol and frequency. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

frequency::String: A time frame.

params...: Additional keyword query params.

See Historical-Stock-Quote for more details.
See Historical-Index-Quote for more details.
See Historical-Euronext-Quote for more details.
See Historical-TSX-Quote for more details.
See Historical-Cryptocurrencies-Quote for more details.
See Historical-Forex-Quote for more details.
See Historical-Commodities-Quote for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the 15m historical price quote for AAPL
data = historical_price_quote(fmp, "AAPL", frequency = TIME_FREQUENCIES.minutes15)

# get the 4hr historical price quote for BTCUSD
data = historical_price_quote(fmp, "BTCUSD", frequency = TIME_FREQUENCIES.hours4)

# get the daily historical price quote time series for EURUSD
data = historical_price_quote(fmp, "EURUSD", frequency = TIME_FREQUENCIES.daily, timeseries = 5)