price_targets

FinancialModelingPrep.price_targets — Function

price_targets(fmp, symbol)

Returns price targets for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Price-Target for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price targets for AAPL
data = price_targets(fmp, "AAPL")

price_targets_by_analyst

FinancialModelingPrep.price_targets_by_analyst — Function

price_targets_by_analyst(fmp, name)

Returns price targets from the specified name.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: An analyst name.

See Price-Target-by-Analyst-Name for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price targets from Tim Anderson
data = price_targets_by_analyst(fmp, name = "Tim%20Anderson")

price_targets_by_company

FinancialModelingPrep.price_targets_by_company — Function

price_targets_by_company(fmp, company)

Returns price targets from the specified company.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

company::String: An analyst company name.

See Price-Target-by-Analyst-Company for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price targets from Barclays
data = price_targets_by_company(fmp, company = "Barclays")

price_targets_summary

FinancialModelingPrep.price_targets_summary — Function

price_targets_summary(fmp, symbol)

Returns the price targets summary for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Price-Target-Summary for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price targets summary for AAPL
data = price_targets_summary(fmp, "AAPL")

price_targets_consensus

FinancialModelingPrep.price_targets_consensus — Function

price_targets_consensus(fmp, symbol)

Returns the price targets consensus for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Price-Target-Consensus for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price targets consensus for AAPL
data = price_target_consensus(fmp, "AAPL")

price_targets_feed

FinancialModelingPrep.price_targets_feed — Function

price_targets_feed(fmp, params...)

Returns a JSON table containing the price targets feed results.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Price-Target-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page from the price target feed
data = price_targets_feed(fmp, page = 0)