crowdfunding_offerings_feed

FinancialModelingPrep.crowdfunding_offerings_feed — Function

crowdfunding_offerings_feed(fmp, params...)

Returns crowdfunding offerings from the RSS feed.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Crowdfunding-Offerings-Rss-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of results from the RSS feed
data = crowdfunding_offerings_feed(fmp, page = 0)

crowdfunding_offerings_search

FinancialModelingPrep.crowdfunding_offerings_search — Function

crowdfunding_offerings_search(fmp, name)

Returns crowfunding offering dates for the specified name.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: A company name.

See Crowdfunding-Offerings-Company-Search for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the crowdfunding offering dates for Enotap
data = crowdfunding_offerings_search(fmp, name = "Enotap")

crowdfunding_offerings

FinancialModelingPrep.crowdfunding_offerings — Function

crowdfunding_offerings(fmp, cik)

Returns the crowdfunding offerings for the specified CIK.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

cik::String: A CIK.

See Crowdfunding-Offerings-by-CIK for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the crowdfundings offerings for OYO Fitness
data = crowdfunding_offerings(fmp, cik = "0001916078")

equity_offerings_feed

FinancialModelingPrep.equity_offerings_feed — Function

equity_offerings_feed(fmp, params...)

Returns equity offerings from the RSS feed.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Equity-Offerings-Fundraising-Rss-feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of results from the RSS feed
data = equity_offerings_feed(fmp, page = 0)

equity_offerings_search

FinancialModelingPrep.equity_offerings_search — Function

equity_offerings_search(fmp, name)

Returns equity offering dates for the specified name.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: A company name.

See Equity-Offerings-Fundraising-Company-Search for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the equity offering dates for Marinalife
data = equity_offerings_search(fmp, name = "Marinalife")

equity_offerings

FinancialModelingPrep.equity_offerings — Function

equity_offerings(fmp, cik)

Returns the crowdfunding offerings for the specified CIK.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

cik::String: A CIK.

See Equity-Offerings-Fundraising-by-CIK for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the equity offerings for Marinalife
data = equity_offerings(fmp, cik = "0001870523")