senate_trades

FinancialModelingPrep.senate_trades — Function

senate_trades(fmp, symbol)

Returns senate trades for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Senate-Trading for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get senate trades for AAPL
data = senate_trades(fmp, "AAPL")

senate_trades_feed

FinancialModelingPrep.senate_trades_feed — Function

senate_trades_feed(fmp, params...)

Returns senate trades from the RSS feed.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Senate-Trading-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of senate trades from the RSS feed
data = senate_trades_feed(fmp, page = 0)

senate_disclosures

FinancialModelingPrep.senate_disclosures — Function

senate_disclosures(fmp, symbol)

Returns senate disclosures for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Senate-Disclosure for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get senate disclosures for AAPL
data = senate_disclosures(fmp, "AAPL")

senate_disclosure_feed

FinancialModelingPrep.senate_disclosure_feed — Function

senate_disclosure_feed(fmp, params...)

Returns senate disclosures from the RSS feed.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Senate-Disclosure-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of senate disclosures from the RSS feed
data = senate_disclosure_feed(fmp, page = 0)