earnings_calendar

FinancialModelingPrep.earnings_calendar — Function

earnings_calendar(fmp, params...)
earnings_calendar(fmp, symbol, params...)

Returns a list of earnings calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Earnings-Calendar for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the earnings calendar events for Q1 2022
data = earnings_calendar(fmp, from = "2022-01-01", to = "2022-03-31")

# get the last 50 earnings calendar events for AAPL
data = earnings_calendar(fmp, "AAPL", limit = 50)

earnings_calendar_confirmed

FinancialModelingPrep.earnings_calendar_confirmed — Function

earnings_calendar_confirmed(fmp, params...)

Returns a list of the confirmed earnings calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Earnings-Calendar-Confirmed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the confirmed earnings calendar events for Q1 2022
data = earnings_calendar_confirmed(fmp, from = "2022-01-01", to = "2022-03-31")

ipo_calendar

FinancialModelingPrep.ipo_calendar — Function

ipo_calendar(fmp, params...)

Returns a list of the ipo calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See IPO-Calendar for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the ipo calendar events for Q1 2022
data = ipo_calendar(fmp, from = "2022-01-01", to = "2022-03-31")

ipo_calendar_prospectus

FinancialModelingPrep.ipo_calendar_prospectus — Function

ipo_calendar_prospectus(fmp, params...)

Returns a list of the ipo calendar events with prospectus. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See IPO-Calendar-with-Prospectus for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the ipo calendar events with prospectus for Q1 2022
data = ipo_calendar_prospectus(fmp, from = "2022-01-01", to = "2022-03-31")

ipo_calendar_confirmed

FinancialModelingPrep.ipo_calendar_confirmed — Function

ipo_calendar_confirmed(fmp, params...)

Returns a list of the confirmed ipo calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See IPO-Calendar-Confirmed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the confirmed ipo calendar events for Q1 2022
data = ipo_calendar_confirmed(fmp, from = "2022-01-01", to = "2022-03-31")

stock_split_calendar

FinancialModelingPrep.stock_split_calendar — Function

stock_split_calendar(fmp, params...)

Returns a list of stock split calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Stock-Split-Calendar for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the stock split calendar events for Q1 2022
data = stock_split_calendar(fmp, from = "2022-01-01", to = "2022-03-31")

dividend_calendar

FinancialModelingPrep.dividend_calendar — Function

dividend_calendar(fmp, params...)

Returns a list of dividend calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Dividend-Calendar for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the dividend calendar events for Q1 2022
data = dividend_calendar(fmp, from = "2022-01-01", to = "2022-03-31")

historical_dividends

FinancialModelingPrep.historical_dividends — Function

historical_dividends(fmp, symbol)

Returns a list historical dividends for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Historical-Dividends for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last historical dividends for AAPL
data = historical_dividends(fmp, "AAPL")

economic_calendar

FinancialModelingPrep.economic_calendar — Function

economic_calendar(fmp, params...)

Returns a list of economic calendar events. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Economic-Calendar for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the economic calendar events for Q1 2022
data = economic_calendar(fmp, from = "2022-01-01", to = "2022-03-31")