symbols_with_financials

FinancialModelingPrep.symbols_with_financials — Function

symbols_with_financials(fmp)

Returns a vector of symbols which have financial statements. Each element is a string.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Financial-Statements-List for more details.

income_statements

FinancialModelingPrep.income_statements — Function

income_statements(fmp, symbol, reported = false, params...)

Returns a vector of income statements for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

reported::Bool: Return the reported or normalized statements.

params...: Additional keyword query params.

See Income-Statements for more details. See Income-Statements-As-Reported for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 10 quarterly statments for AAPL
data = income_statements(fmp, "AAPL", period = "quarter", limit = 10)

# get the last 5 annual statements as reported for AAPL
data = income_statements(fmp, "AAPL", reported = true, limit = 5)

balance_sheet_statements

FinancialModelingPrep.balance_sheet_statements — Function

balance_sheet_statements(fmp, symbol, reported = false, params...)

Returns a vector of balance sheet statements for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

reported::Bool: Return the reported or normalized statements.

params...: Additional keyword query params.

See Balance-Sheet-Statements for more details. See Balance-Sheet-Statements-As-Reported for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 10 quarterly statments for AAPL
data = balance_sheet_statements(fmp, "AAPL", period = "quarter", limit = 10)

# get the last 5 annual statements as reported for AAPL
data = balance_sheet_statements(fmp, "AAPL", reported = true, limit = 5)

cash_flow_statements

FinancialModelingPrep.cash_flow_statements — Function

cash_flow_statements(fmp, symbol, reported = false, params...)

Returns a vector of cash flow statements for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

reported::Bool: Return the reported or normalized statements.

params...: Additional keyword query params.

See Cash-Flow-Statements for more details. See Cash-Flow-Statements-As-Reported for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 10 quarterly statments for AAPL
data = cash_flow_statements(fmp, "AAPL", period = "quarter", limit = 10)

# get the last 5 annual statements as reported for AAPL
data = cash_flow_statements(fmp, "AAPL", reported = true, limit = 5)

financial_statements

FinancialModelingPrep.financial_statements — Function

financial_statements(fmp, symbol, params...)

Returns a vector of financial statements as reported for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Full-Financial-Statements-As-Reported for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all quarterly statements as reported for AAPL
data = financial_statements(fmp, "AAPL", period = "quarter")

financial_reports

FinancialModelingPrep.financial_reports — Function

financial_reports(fmp, symbol, year, period = "FY")

Returns a vector with the financial report for the specified symbol, year and period. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

year::Integer: A calendar year.

period::String: One of "FY", "Q1", "Q2", "Q3" or "Q4".

See Annual-Reports-on-Form-10-K for more details. See Quarterly-Earnings-Reports for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the 10-K for AAPL in 2022
data = financial_reports(fmp, "AAPL", year = 2022)

# get the 10-Q for AAPL in Q4 of 2022
data = financial_reports(fmp, "AAPL", year = 2022, period = "Q4")

revenue_segments

FinancialModelingPrep.revenue_segments — Function

revenue_segments(fmp, symbol, segment = REVENUE_SEGMENTS.product, params...)

Returns a dictionary with the revenue segments for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

segment::String: A REVENUE_SEGMENTS option.

params...: Additional keyword query params.

See Sales-Revenue-By-Segments for more details. See Revenue-Geographic-by-Segments for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all yearly geographic revenue segments for AAPL
data = revenue_segments(fmp, "AAPL", segment = REVENUE_SEGMENTS.geographic)

# get all quarterly product revenue segments for AAPL
data = revenue_segments(fmp, "AAPL", segment = REVENUE_SEGMENTS.product, period = "quarter")

shares_float

FinancialModelingPrep.shares_float — Function

shares_float(fmp, symbol = "all")

Returns a vector of shares float statistics for one or all symbols. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol or all symbols if not provided.

See Shares-Float for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get shares float for all symbols
data = shares_float(fmp)

# get shares float for AAPL
data = shares_float(fmp, "AAPL")

earnings_call_transcripts

FinancialModelingPrep.earnings_call_transcripts — Function

earnings_call_transcript(fmp, symbol)

Returns a vector of earnings call transcripts for a specified symbols. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol or "all" if not provided.

params...: Additional keyword query params.

See Earnings-Call-Transcript for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all transcripts for AAPL in Q3 of 2022
data = earnings_call_transcript(fmp, "AAPL", year = 2022, quarter = 3)

sec_filings

FinancialModelingPrep.sec_filings — Function

sec_filings(fmp, symbol)

Returns a vector of sec filings for a specified symbols. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol or "all" if not provided.

params...: Additional keyword query params.

See SEC-Filings for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all 10-K filings for AAPL on page 2
data = sec_filings(fmp, "AAPL", type = "10-K", page = 2)

company_notes

FinancialModelingPrep.company_notes — Function

company_notes(fmp, symbol)

Returns a vector of notes due for a specified symbols. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Company-Notes-Due for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all notes due for AAPL
data = company_notes(fmp, "AAPL")