financial_ratios

FinancialModelingPrep.financial_ratios — Function

financial_ratios(fmp, symbol, params...)

Returns a vector of common financial ratios for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Financial-Ratios for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get financial ratios for AAPL in the last 30 quarters
data = financial_ratios(fmp, "AAPL", period = "quarter", limit = 30)

financial_scores

FinancialModelingPrep.financial_scores — Function

financial_scores(fmp, symbol)

Returns a vector of common financial scores for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Financial-Scores for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get financial scores for AAPL
data = financial_scores(fmp, "AAPL")

owners_earnings

FinancialModelingPrep.owners_earnings — Function

owners_earnings(fmp, symbol)

Returns a vector of owners earnings for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Owners-Earnings for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get ownings earnings for AAPL
data = owners_earnings(fmp, "AAPL")

enterprise_values

FinancialModelingPrep.enterprise_values — Function

enterprise_values(fmp, symbol, params...)

Returns a vector of enterprise value components for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Enterprise-Value for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get enterprise values for AAPL in the last 30 quarters
data = enterprise_values(fmp, "AAPL", period = "quarter", limit = 30)

income_statements_growth

FinancialModelingPrep.income_statements_growth — Function

income_statements_growth(fmp, symbol, params...)

Returns a vector of income statements growth for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Income-Statements-Growth for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 5 annual statements growth for AAPL
data = income_statements_growth(fmp, "AAPL", limit = 5)

balance_sheet_statements_growth

FinancialModelingPrep.balance_sheet_statements_growth — Function

balance_sheet_statements_growth(fmp, symbol, params...)

Returns a vector of balance sheet statements growth for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Balance-Sheet-Statements-Growth for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 5 annual statements growth for AAPL
data = balance_sheet_statements_growth(fmp, "AAPL", limit = 5)

cash_flow_statements_growth

FinancialModelingPrep.cash_flow_statements_growth — Function

cash_flow_statements_growth(fmp, symbol, params...)

Returns a vector of cash flow statements growth for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Cash-Flow-Statements-Growth for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 5 annual statements growth for AAPL
data = cash_flow_statements_growth(fmp, "AAPL", limit = 5)

financial_statements_growth

FinancialModelingPrep.financial_statements_growth — Function

financial_statements_growth(fmp, symbol, params...)

Returns a vector of financial statements growth for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Financial-Statements-Growth for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 5 annual statements growth for AAPL
data = financial_statements_growth(fmp, "AAPL", limit = 5)

key_metrics

FinancialModelingPrep.key_metrics — Function

key_metrics(fmp, symbol, period = REPORTING_PERIODS.ttm, params...)

Returns a vector of key metrics for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

period::String: A REPORTING_PERIODS option.

params...: Additional keyword query params.

See Key-Metrics for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get key metrics for AAPL in the last 30 years by ttm
data = key_metrics(fmp, "AAPL", period = REPORTING_PERIODS.ttm, limit = 30)

company_rating

FinancialModelingPrep.company_rating — Function

company_rating(fmp, symbol)

Returns a vector of ratings for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Company-Rating for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the company rating for AAPL 
data = company_rating(fmp, "AAPL")

historical_ratings

FinancialModelingPrep.historical_ratings — Function

historical_ratings(fmp, symbol, params...)

Returns a vector of ratings for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Historical-Ratings for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 100 ratings for AAPL 
data = historical_ratings(fmp, "AAPL", limit = 100)

discounted_cash_flows

FinancialModelingPrep.discounted_cash_flows — Function

discounted_cash_flows(fmp, symbol, with_wacc = false)

Returns a vector of discounted cash flows for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

with_wacc::Bool: Return the version with WACC.

See Discounted-Cash-Flow for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the latest dcf for AAPL including WACC
data = discounted_cash_flows(fmp, "AAPL", with_wacc = true)

historical_discounted_cash_flows

FinancialModelingPrep.historical_discounted_cash_flows — Function

historical_discounted_cash_flows(fmp, symbol, params...)

Returns a vector of historical discounted cash flows for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Historical-Discounted-Cash-Flow for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the dcf for in the last 30 quarters AAPL
data = historical_discounted_cash_flows(fmp, "AAPL", period = "quarter", limit = 30)