available_symbols

FinancialModelingPrep.available_symbols — Function

available_symbols(fmp)

Returns a vector of all available symbols in the API. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Symbols-List for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get a list of all available symbols
data = available_symbols(fmp)

tradeable_symbols

FinancialModelingPrep.tradeable_symbols — Function

tradeable_symbols(fmp)

Returns a vector of all tradeable symbols in the API. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Tradeable-Symbols-List for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get a list of all available symbols
data = tradeable_symbols(fmp)

etf_symbols

FinancialModelingPrep.etf_symbols — Function

etf_symbols(fmp)

Returns a vector of all tradeable symbols in the API. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See ETF-Symbols for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get a list of all etf symbols
data = etf_symbols(fmp)