otc_quote

FinancialModelingPrep.otc_quote — Function

otc_quote(fmp, symbol)

Returns the price quote for the specified symbol(s).

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See OTC-Quote for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the otc quote for GBTC
data = otc_quote(fmp, "GBTC")

price_change

FinancialModelingPrep.price_change — Function

price_change(fmp, symbol)

Returns the price change for the specified symbol(s).

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Price-Change for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the price change for AAPL
data = price_change(fmp, "AAPL")

historical_splits

FinancialModelingPrep.historical_splits — Function

historical_splits(fmp, symbol)

Returns the historical stock splits for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Historical-Stock-Splits for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the historical splits for AAPL
data = historical_splits(fmp, "AAPL")

survivorship_bias

FinancialModelingPrep.survivorship_bias — Function

survivorship_bias(fmp, symbol, date)

Returns a JSON dictionary of the survivorship bias for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

date::String: A yyyy-mm-dd formatted date string.

See Survivorship-Bias for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the survivorship bias for AAPL to Jan 3, 2012
data = survivorship_bias(fmp, "AAPL", date = "2012-01-03")

technical_indicators

FinancialModelingPrep.technical_indicators — Function

technical_indicators(fmp, symbol, frequency = TIME_FREQUENCIES.daily, period = 200, type = "SMA")

Returns the historical price quote for the specified symbol and frequency.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

frequency::String: A time frequency.

period::Integer: The indicator period.

type::String: The indicator type.

See Daily-Indicators for more details.
See Intraday-Indicators for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the daily 50 period SMA for AAPL
data = technical_indicators(fmp, "AAPL", period = 50)

# get the 15m 10 period WMA for AAPL
data = technical_indicators(fmp, "AAPL", TIME_FREQUENCIES.minutes15, period = 10, type = "wma")