search_symbol

FinancialModelingPrep.search_symbol — Function

search_symbol(fmp, symbol, params...)

Returns search results for the specified symbol and filters.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Symbol-Search for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first 10 matches where symbol is AA and exchange is NASDAQ
data = search_symbol(fmp, "AA", limit = 10, exchange = "NASDAQ")

search_name

FinancialModelingPrep.search_name — Function

search_name(fmp, name, params...)

Returns search results for the specified name and filters.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: A company name.

params...: Additional keyword query params.

See Name-Search for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first 10 matches where name is Meta and exchange is NASDAQ
data = search_name(fmp, "Meta", limit = 10, exchange = "NASDAQ")

stock_screener

FinancialModelingPrep.stock_screener — Function

stock_screener(fmp, params...)

Returns search results for the specified filters.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Stock-Screener for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all tech stocks in the NASDAQ with a market cap over 100mm and a beta above 1
data = stock_screener(fmp, marketCapMoreThan = 100000000, betaMoreThan = 1, sector = "Technology", exchange = "NASDAQ")

# get the first 100 stocks in California with a price greater than 100
data = stock_screener(fmp, country = "CA", priceMoreThan = 100, limit = 100)

available_countries

FinancialModelingPrep.available_countries — Function

available_countries(fmp)

Returns a JSON array of all available countries in the API.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Stock-Screener for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get a list of all available countries
data = available_countries(fmp)