historical_social_sentiment

FinancialModelingPrep.historical_social_sentiment — Function

historical_social_sentiment(fmp, symbol, params...)

Returns a JSON table with the historical social sentiment for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Social-Sentiment for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of historical social sentiment for AAPL
data = historical_social_sentiment(fmp, "AAPL", page = 0)

social_sentiment_trends

FinancialModelingPrep.social_sentiment_trends — Function

social_sentiment_trends(fmp, type, source)

Returns a JSON table with the social sentiment trends for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

type::String: One of "bullish" or "bearish".

source::String: One of "twitter" or "stocktwits".

See Social-Sentiment for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all social sentiment trends
data = social_sentiment_trends(fmp)

# get all social sentiment trends from stocktwits with a bearish bias
data = social_sentiment_trends(fmp, type = "bearish", source = "stocktwits")

social_sentiment_changes

FinancialModelingPrep.social_sentiment_changes — Function

social_sentiment_changes(fmp, type, source)

Returns a JSON table with the social sentiment changes for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

type::String: One of "bullish" or "bearish".

source::String: One of "twitter" or "stocktwits".

See Social-Sentiment for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all social sentiment changes
data = social_sentiment_trends(fmp)

# get all social sentiment changes from stocktwits with a bearish bias
data = social_sentiment_changes(fmp, type = "bearish", source = "stocktwits")

stock_grades

FinancialModelingPrep.stock_grades — Function

stock_grades(fmp, symbol, params...)

Returns stock grades for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Stock-Grade for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 100 stock grades for AAPL
data = stock_grades(fmp, "AAPL", limit = 100)

earnings_surprises

FinancialModelingPrep.earnings_surprises — Function

earnings_surprises(fmp, symbol)

Returns earnings suprises for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Earnings-Surprises for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the latest earnings surprises for AAPL
data = earnings_surprises(fmp, "AAPL")

analyst_estimates

FinancialModelingPrep.analyst_estimates — Function

analyst_estimates(fmp, symbol, params...)

Returns analyst estimates for the specified symbol.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Analyst-Estimates for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 4 quarters of analyst estimates for AAPL
data = analyst_estimates(fmp, "AAPL", period = "quarter", limit = 4)

mergers_and_acquisitions_feed

FinancialModelingPrep.mergers_and_acquisitions_feed — Function

mergers_and_acquisitions_feed(fmp, params...)

Returns a JSON table containing the mergers and acquisitions feed results.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Mergers-and-Acquisitions-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of the mergers and acquisitions rss feed
data = mergers_and_acquisitions_feed(fmp, page = 0)

mergers_and_acquisitions_search

FinancialModelingPrep.mergers_and_acquisitions_search — Function

mergers_and_acquisitions_search(fmp, params...)

Returns a JSON table containing mergers and acquisitions search results.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Mergers-and-Acquisitions for more details.

Examples

# create a FMP API instance
fmp = FMP()

# search for mergers and acquisitions
data = mergers_and_acquisitions_search(fmp, name = "Syros")