insider_trading_types

FinancialModelingPrep.insider_trading_types — Function

insider_trading_types(fmp)

Returns a vector of insider trading transaction types. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

See Insider-Trading for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the types of insider trading events
data = insider_trading_types(fmp)

insider_trades

FinancialModelingPrep.insider_trades — Function

insider_trades(fmp, params...)

Returns a vector of insider trades. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Insider-Trading for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the most recent insider purchases and sales
data = insider_trades(fmp, transactionType = ["P-Purchase", "S-Sale"], page = 0)

# get the most recent insider trades for AAPL
data = insider_trades(fmp, symbol = "AAPL", page = 0)

insider_trades_feed

FinancialModelingPrep.insider_trades_feed — Function

insider_trades_feed(fmp, params...)

Returns a vector of insider trades from the RSS feed. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Insider-Trading-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the most recent insider trades from the feed
data = insider_trades_feed(fmp, page = 0)

cik_list

FinancialModelingPrep.cik_list — Function

cik_list(fmp, params...)

Returns a list of all CIKs matching the specified parameters. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See CIK-Mapper for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all CIKs from page 3
data = cik_list(fmp, page = 3)

cik_from_name

FinancialModelingPrep.cik_from_name — Function

cik_from_name(fmp, name)

Returns a list of all CIKs matching the specified name. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

name::String: A name.

See CIK-Mapper for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get all CIKs matching the name
data = cik_from_name(fmp, "zuckerberg%20mark")

cik_from_symbol

FinancialModelingPrep.cik_from_symbol — Function

cik_from_symbol(fmp, symbol)

Returns a list of all CIKs matching the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::symbol: A stock symbol.

See CIK-Mapper for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the CIK for AAPL
data = cik_from_symbol(fmp, "AAPL")

insider_roster

FinancialModelingPrep.insider_roster — Function

insider_roster(fmp, symbol)

Returns a list with the insider roster for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::symbol: A stock symbol.

See Insider-Roster for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the insider roster for AAPL
data = insider_roster(fmp, "AAPL")

insider_roster_statistics

FinancialModelingPrep.insider_roster_statistics — Function

insider_roster_statistics(fmp, symbol)

Returns a list of insider roster statistics for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::symbol: A stock symbol.

See Insider-Roster-Statistics for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the insider roster statistics for AAPL
data = insider_roster_statistics(fmp, "AAPL")

fails_to_deliver

FinancialModelingPrep.fails_to_deliver — Function

fails_to_deliver(fmp, symbol, params)

Returns a list of fails to deliver for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::symbol: A stock symbol.

params...: Additional keyword query params.

See Fails-to-Deliver for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of fails to deliver for AAPL
data = fails_to_deliver(fmp, "AAPL", page = 0)