social_sentiment

FinancialModelingPrep.social_sentiment — Function

social_sentiment(fmp, symbol, params...)
social_sentiment(fmp, type = "trending", params...)

Returns a vector with the social sentiment type for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

type::String: One of "trending", or "change".

params...: Additional keyword query params.

See Social-Sentiment for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of historical social sentiment for AAPL
data = social_sentiment(fmp, "AAPL", page = 0)

# get all trending social sentiment from stocktwits with a bullish bias
data = social_sentiment(fmp, type = "trending", type = "bullish", source = "stocktwits")

stock_grades

FinancialModelingPrep.stock_grades — Function

stock_grades(fmp, symbol, params...)

Returns a vector of stock grades for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Stock-Grade for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 100 stock grades for AAPL
data = stock_grades(fmp, "AAPL", limit = 100)

earnings_surprises

FinancialModelingPrep.earnings_surprises — Function

earnings_surprises(fmp, symbol)

Returns a vector of earnings suprises for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

See Earnings-Surprises for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the latest earnings surprises for AAPL
data = earnings_surprises(fmp, "AAPL")

analyst_estimates

FinancialModelingPrep.analyst_estimates — Function

analyst_estimates(fmp, symbol, params...)

Returns a vector of analyst estimates for the specified symbol. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

symbol::String: A stock symbol.

params...: Additional keyword query params.

See Analyst-Estimates for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the last 4 quarters of analyst estimates for AAPL
data = analyst_estimates(fmp, "AAPL", period = "quarter", limit = 4)

mergers_and_acquisitions_feed

FinancialModelingPrep.mergers_and_acquisitions_feed — Function

mergers_and_acquisitions_feed(fmp, params...)

Returns a vector containing the mergers and acquisitions feed results. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Mergers-and-Acquisitions-RSS-Feed for more details.

Examples

# create a FMP API instance
fmp = FMP()

# get the first page of the mergers and acquisitions rss feed
data = mergers_and_acquisitions_feed(fmp, page = 0)

mergers_and_acquisitions_search

FinancialModelingPrep.mergers_and_acquisitions_search — Function

mergers_and_acquisitions_search(fmp, params...)

Returns a vector containing mergers and acquisitions search results. Each element is a dictionary.

Arguments

fmp::FMP: A Financial Modeling Prep instance.

params...: Additional keyword query params.

See Mergers-and-Acquisitions for more details.

Examples

# create a FMP API instance
fmp = FMP()

# search for mergers and acquisitions
data = mergers_and_acquisitions_search(fmp, name = "Syros")